测试27更新
1. 同步官版更新

2. 去除了一些导致编译器警告的

3. 离子炮随机改用游戏提供的Random

4. 修正自定义经验的问题

5. 专用化ValueableVector<bool>的存读档，受击武器的ValueableVector<bool>恢复

6. YRpp

+++ src/Ext/WeaponType/Body.cpp
@@ -104,7 +104,7 @@ void WeaponTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->ElectricLaser_Count.Read(exINI, pSection, "ElectricLaser.Count");
 	this->ElectricLaser_Length.Read(exINI, pSection, "ElectricLaser.Length");
 
-	for (size_t i = 0; i < this->ElectricLaser_Count; ++i)
+	for (int i = 0; i < this->ElectricLaser_Count; ++i)
 	{
 		char colortext[64];
 		Nullable<ColorStruct> color;
